converted tests for  GT-11138 tp enyo 2.6
