+++ test/moonstone/ProgressBar/GT-11138-ProgressButtonDigitAnimation/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "GT-11138-ProgressButtonDigitAnimation.js"
+}
